refactor project service

+++ app/projects/controller.py
@@ -1,12 +1,11 @@
 import datetime
 import json
 import os
-import re
 import base64
 from typing import List
 
 
-from flask import abort, current_app, request, jsonify
+from flask import abort, current_app, request
 from flask_accepts.decorators.decorators import accepts, responds
 from flask_login import current_user
 from flask_restx import Namespace, Resource, reqparse
@@ -14,29 +13,14 @@ from werkzeug.utils import secure_filename
 import werkzeug
 
 from app.utils.grew_utils import GrewService
-from .interface import (
-    ProjectExtendedInterface,
-    ProjectInterface,
-    ProjectShownFeaturesAndMetaInterface,
-)
-from .model import Project, ProjectAccess
-from .schema import (
-    ProjectExtendedSchema,
-    ProjectSchema,
-    ProjectSchemaCamel,
-    ProjectFeaturesAndMetaSchema,
-)
-from .service import (
-    LastAccessService,
-    ProjectAccessService,
-    ProjectFeatureService,
-    ProjectMetaFeatureService,
-    ProjectService,
-)
 from app.user.service import UserService
 
-api = Namespace("Project", description="Endpoints for dealing with projects")  # noqa
+from .interface import ProjectExtendedInterface, ProjectInterface, ProjectShownFeaturesAndMetaInterface
+from .model import Project, ProjectAccess
+from .schema import ProjectExtendedSchema, ProjectSchema, ProjectFeaturesAndMetaSchema
+from .service import LastAccessService, ProjectAccessService, ProjectFeatureService, ProjectMetaFeatureService, ProjectService
 
+api = Namespace("Project", description="Endpoints for dealing with projects")  # noqa
 
 @api.route("/")
 class ProjectResource(Resource):
@@ -52,9 +36,9 @@ class ProjectResource(Resource):
 
         grew_projects = GrewService.get_projects()
 
-        grewnames = set([project["name"] for project in grew_projects])
-        dbnames = set([project.project_name for project in projects])
-        common = grewnames & dbnames
+        grew_projects_names = set([project["name"] for project in grew_projects])
+        db_projects_names = set([project.project_name for project in projects])
+        common = grew_projects_names & db_projects_names
 
         for project in projects:
             if ProjectAccessService.check_project_access(
@@ -70,7 +54,6 @@ class ProjectResource(Resource):
                     project.guests,
                 ) = ProjectAccessService.get_all(project.id)
 
-                # better version: less complexity + database calls / 2
                 (
                     last_access,
                     last_write_access,
@@ -81,6 +64,14 @@ class ProjectResource(Resource):
                 project.last_access = last_access - now
                 project.last_write_access = last_write_access - now
 
+                project_image = project.image
+                if project_image:
+                    image_path = os.path.join(current_app.config["PROJECT_IMAGE_FOLDER"], project_image)
+                    if os.path.exists(image_path):
+                        with open(image_path, 'rb') as file:
+                            image_data = base64.b64encode(file.read()).decode('utf-8')
+                            project.image = 'data:image/${};base64,${}'.format(image_data, image_path.split(".")[1])
+        
                 for grew_project in grew_projects:
                     if grew_project["name"] == project.project_name:
                         project.users = grew_project["users"]
@@ -91,43 +82,24 @@ class ProjectResource(Resource):
                 projects_extended_list.append(project)
 
         return projects_extended_list
-
+    
+    @accepts(schema=ProjectSchema)
     @responds(schema=ProjectSchema)
     def post(self) -> Project:
-        "Create a single Project"
+        """Create a single Project"""
+        
         try:
             creator_id = current_user.id
         except:
             abort(401, "User not loged in")
+        
+        new_project_attrs: ProjectInterface = request.parsed_obj
 
-        # KK : Make a unified schema for all http request related to project
-        # ... and have the schema taking JS camelcase typing
-        parser = reqparse.RequestParser()
-        parser.add_argument(name="projectName", type=str)
-        parser.add_argument(name="description", type=str)
-        parser.add_argument(name="blindAnnotationMode", type=bool)
-        parser.add_argument(name="visibility", type=int)
-        parser.add_argument(name="config", type=str)
-        parser.add_argument(name="language", type=str)
-        args = parser.parse_args()
-        projectName =  args.projectName
-
-        new_project_attrs: ProjectInterface = {
-            "project_name": projectName,
-            "description": args.description,
-            "blind_annotation_mode": args.blindAnnotationMode,
-            "visibility": args.visibility,
-            "freezed": False,
-            "config": args.config,
-            "language": args.language
-        }
-
-        # KK : TODO : put all grew request in a seperated file and add error catching
         GrewService.create_project(new_project_attrs["project_name"])
-
         new_project = ProjectService.create(new_project_attrs)
+        
         LastAccessService.update_last_access_per_user_and_project(
-            creator_id, projectName, "write"
+            creator_id, new_project_attrs["project_name"], "write"
         )
 
         ProjectAccessService.create(
@@ -142,48 +114,46 @@ class ProjectResource(Resource):
         default_metafeatures = ["text_en"]
 
         for feature in default_features:
-            ProjectFeatureService.create(
-                {"project_id": new_project.id, "value": feature}
-            )
+            ProjectFeatureService.create({"project_id": new_project.id, "value": feature})
 
         for feature in default_metafeatures:
-            ProjectMetaFeatureService.create(
-                {"project_id": new_project.id, "value": feature}
-            )
+            ProjectMetaFeatureService.create({"project_id": new_project.id, "value": feature})
 
         return new_project
 
 
-@api.route("/<string:projectName>")
+@api.route("/<string:project_name>")
 class ProjectIdResource(Resource):
-    """Views for dealing with single identified project"""
 
-    @responds(schema=ProjectSchemaCamel, api=api)
-    def get(self, projectName: str):
+    @responds(schema=ProjectSchema, api=api)
+    def get(self, project_name: str):
         """Get a single project"""
-        return ProjectService.get_by_name(projectName)
+        
+        return ProjectService.get_by_name(project_name)
 
-    @responds(schema=ProjectSchemaCamel, api=api)
-    @accepts(schema=ProjectSchemaCamel, api=api)
-    def put(self, projectName: str):
-        """Modify a single project (by it's name)"""
-        project = ProjectService.get_by_name(projectName)
+    @responds(schema=ProjectSchema, api=api)
+    @accepts(schema=ProjectSchema, api=api)
+    def put(self, project_name: str):
+        """Modify a project (by it's name)"""
+        
+        project = ProjectService.get_by_name(project_name)
         ProjectAccessService.check_admin_access(project.id)
+        
         changes: ProjectInterface = request.parsed_obj
-        if ('project_name' in changes.keys()):
-            ProjectService.rename_project(projectName, changes.get("project_name"))
-            
+        if 'project_name' in changes.keys():
+            GrewService.rename_project(project_name, changes.get("project_name"))  
         return ProjectService.update(project, changes)
 
-    def delete(self, projectName: str):
-        """Delete a single project (by it's name)"""
-        project = ProjectService.get_by_name(projectName)
+    def delete(self, project_name: str):
+        """Delete a project (by it's name)"""
+        
+        project = ProjectService.get_by_name(project_name)
         ProjectAccessService.check_admin_access(project.id)
 
-        project_name = ProjectService.delete_by_name(projectName)
+        project_name = ProjectService.delete_by_name(project_name)
         if project_name:
             GrewService.delete_project(project_name)
-            return {"status": "Success", "projectName": project_name}
+            return {"status": "Success", "project_name": project_name}
         else:
             return {
                 "status": "Error",
@@ -193,12 +163,15 @@ class ProjectIdResource(Resource):
             }
 
 
-@api.route("/<string:projectName>/features")
+@api.route("/<string:project_name>/features")
 class ProjectFeaturesResource(Resource):
+    """Class for dealing with project features"""
+    
     @responds(schema=ProjectFeaturesAndMetaSchema, api=api)
-    def get(self, projectName: str):
-        """Get a single project features"""
-        project = ProjectService.get_by_name(projectName)
+    def get(self, project_name: str):
+        """Get project features"""
+        
+        project = ProjectService.get_by_name(project_name)
         ProjectService.check_if_project_exist(project)
 
         features = {
@@ -208,19 +181,22 @@ class ProjectFeaturesResource(Resource):
         return features
 
     @accepts(schema=ProjectFeaturesAndMetaSchema, api=api)
-    def put(self, projectName: str):
-        parsed_obj: ProjectShownFeaturesAndMetaInterface = request.parsed_obj
-        project = ProjectService.get_by_name(projectName)
+    def put(self, project_name: str):
+        """Update project features"""
+        
+        project = ProjectService.get_by_name(project_name)
+        ProjectService.check_if_project_exist(project)
         ProjectAccessService.check_admin_access(project.id)
-
-        shown_features = parsed_obj.get("shown_features")
+        
+        new_shown_features_shown_meta: ProjectShownFeaturesAndMetaInterface = request.parsed_obj
+        shown_features = new_shown_features_shown_meta["shown_features"]
         if shown_features is not None and isinstance(shown_features, list):
             ProjectFeatureService.delete_by_project_id(project.id)
             for feature in shown_features:
                 new_attrs = {"project_id": project.id, "value": feature}
                 ProjectFeatureService.create(new_attrs)
 
-        shown_meta = parsed_obj.get("shown_meta")
+        shown_meta = new_shown_features_shown_meta["shown_features"]
         if shown_meta is not None and isinstance(shown_meta, list):
             ProjectMetaFeatureService.delete_by_project_id(project.id)
             for feature in shown_meta:
@@ -230,51 +206,54 @@ class ProjectFeaturesResource(Resource):
         return {"status": "success"}
 
 
-@api.route("/<string:projectName>/conll-schema")
+@api.route("/<string:project_name>/conll-schema")
 class ProjectConllSchemaResource(Resource):
-    def get(self, projectName: str):
-        """Get a single project conll schema"""
-        project = ProjectService.get_by_name(projectName)
+    
+    def get(self, project_name: str):
+        """Get project conll schema"""
+        
+        project = ProjectService.get_by_name(project_name)
         ProjectService.check_if_project_exist(project)
-        conll_schema = GrewService.get_conll_schema(projectName)
-        return conll_schema
+        return GrewService.get_conll_schema(project_name)
 
-    def put(self, projectName: str):
-        """Modify a single project conll schema"""
-        project = ProjectService.get_by_name(projectName)
+    def put(self, project_name: str):
+        """Modify project conll schema"""
+        
+        project = ProjectService.get_by_name(project_name)
         ProjectService.check_if_project_exist(project)
         ProjectAccessService.check_admin_access(project.id)
-        parser = reqparse.RequestParser()
-        parser.add_argument(name="config", type=dict, action="append")
-        args = parser.parse_args()
+        
+        args = request.get_json()
         dumped_project_config = json.dumps(args.config)
         GrewService.update_project_config(project.project_name, dumped_project_config)
 
         return {"status": "success", "message": "New conllu schema was saved"}
 
 
-@api.route("/<string:projectName>/access")
+@api.route("/<string:project_name>/access")
 class ProjectAccessResource(Resource):
-    def get(self, projectName: str):
-        """Get a single project users access"""
-        project = ProjectService.get_by_name(projectName)
+    
+    def get(self, project_name: str):
+        """Get project users access"""
+        
+        project = ProjectService.get_by_name(project_name)
         ProjectService.check_if_project_exist(project)
         return ProjectAccessService.get_users_role(project.id)
 
 
-@api.route("/<string:projectName>/access/many")
+@api.route("/<string:project_name>/access/many")
 class ProjectAccessManyResource(Resource):
-    def put(self, projectName: str):
-        """Modify a single project users access"""
-        parser = reqparse.RequestParser()
-        parser.add_argument(name="selectedUsers", type=str, action="append")
-        parser.add_argument(name="targetRole", type=str)
-        args = parser.parse_args()
-        selected_users = args.get("selectedUsers")
-        target_role = args.get("targetRole")
+    
+    def put(self, project_name: str):
+        """Modify project users access"""
+        
+        args = request.get_json()
+        selected_users = args["selectedUsers"]
+        target_role = args["targetRole"]
 
-        project = ProjectService.get_by_name(projectName)
+        project = ProjectService.get_by_name(project_name)
         ProjectService.check_if_project_exist(project)
+        ProjectAccessService.check_admin_access(project.id)
         
         access_level = ProjectAccess.LABEL_TO_LEVEL[target_role]
 
@@ -288,61 +267,49 @@ class ProjectAccessManyResource(Resource):
             }
             project_access = ProjectAccessService.get_by_user_id(user_id, project.id)
             if project_access:
-                project_access = ProjectAccessService.update(project_access, new_attrs)
+                ProjectAccessService.update(project_access, new_attrs)
             else:
-                project_access = ProjectAccessService.create(new_attrs)
+                ProjectAccessService.create(new_attrs)
 
         return ProjectAccessService.get_users_role(project.id)
 
 
-@api.route("/<string:projectName>/access/<string:username>")
+@api.route("/<string:project_name>/access/<string:username>")
 class ProjectAccessUserResource(Resource):
-    def delete(self, projectName: str, username: str):
-        project = ProjectService.get_by_name(projectName)
+    
+    def delete(self, project_name: str, username: str):
+        """Remove project user access"""
+        
+        project = ProjectService.get_by_name(project_name)
         ProjectService.check_if_project_exist(project)
+        ProjectAccessService.check_admin_access(project.id)
+        
         user_id = UserService.get_by_username(username).id
         ProjectAccessService.delete(user_id, project.id)
         return ProjectAccessService.get_users_role(project.id)
 
 
-@api.route("/<string:projectName>/image")
+@api.route("/<string:project_name>/image")
 class ProjectImageResource(Resource):
-
-    def get(self, projectName: str):
-        project = ProjectService.get_by_name(projectName)
-        ProjectService.check_if_project_exist(projectName)
-
-        project_image = project.image
-        if(project_image):
-            image_path = os.path.join(current_app.config["PROJECT_IMAGE_FOLDER"], project_image)
-            if os.path.exists(image_path):
-                with open(image_path, 'rb') as file:
-                    image_data = base64.b64encode(file.read()).decode('utf-8')
-                    return jsonify({"image_data": image_data, "image_ext": image_path.split(".")[1]})
-        return jsonify({})
     
-
-    @responds(schema=ProjectSchemaCamel)
-    def post(self, projectName: str):
+    def post(self, project_name: str):
+        """Update project image"""
+        
         parser = reqparse.RequestParser()
-        parser.add_argument(
-            "files", type=werkzeug.datastructures.FileStorage, location="files"
-        )
+        parser.add_argument("files", type=werkzeug.datastructures.FileStorage, location="files")
         args = parser.parse_args()
-        project = ProjectService.get_by_name(projectName)
+        
+        project = ProjectService.get_by_name(project_name)
         ProjectService.check_if_project_exist(project)
-
-        file_ext = os.path.splitext(args["files"].filename)[1]
-        if file_ext not in current_app.config["UPLOAD_IMAGE_EXTENSIONS"]:
+        ProjectAccessService.check_admin_access(project.id)
+         
+        file_extension = os.path.splitext(args["files"].filename)[1]
+        if file_extension not in current_app.config["UPLOAD_IMAGE_EXTENSIONS"]:
             abort(400)
 
-        filename = secure_filename(projectName + file_ext)
+        file_name = secure_filename(project_name + file_extension)
         content = args["files"].read()
-        with open(
-            os.path.join(current_app.config["PROJECT_IMAGE_FOLDER"], filename), "wb"
-        ) as f:
+        with open(os.path.join(current_app.config["PROJECT_IMAGE_FOLDER"], file_name), "wb") as f:
             f.write(content)
-
-        ProjectService.change_image(projectName, filename)
-
-        return ProjectService.get_by_name(projectName)
+        ProjectService.update(project, {"image": file_name})
+
